requirements.txt
langchain
pdfplumber==0.10.1
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5
torch==2.0.1  # Needed for sentence-transformers
pickle5==0.0.11  # For handling pickle files
transformers==4.12.3
nltk==3.6.7
Pillow
streamlit
python-dotenv
sentence-transformers